Starcast is 60+% faster

+++ src/main/java/com/volmit/iris/engine/actuator/IrisDecorantActuator.java
@@ -78,14 +78,14 @@ public class IrisDecorantActuator extends EngineAssignedActuator<BlockData> {
             int finalI = i;
             burst.queue(() -> {
                 int height;
-                int realX = (int) Math.round(x + finalI);
+                int realX = Math.round(x + finalI);
                 int realZ;
                 IrisBiome biome, cave;
                 for (int j = 0; j < output.getDepth(); j++) {
                     boolean solid;
                     int emptyFor = 0;
                     int lastSolid = 0;
-                    realZ = (int) Math.round(z + j);
+                    realZ = Math.round(z + j);
                     height = (int) Math.round(getComplex().getHeightStream().get(realX, realZ));
                     biome = getComplex().getTrueBiomeStream().get(realX, realZ);
                     cave = shouldRay ? getComplex().getCaveBiomeStream().get(realX, realZ) : null;
@@ -96,8 +96,8 @@ public class IrisDecorantActuator extends EngineAssignedActuator<BlockData> {
 
                     if (height < getDimension().getFluidHeight()) {
                         getSeaSurfaceDecorator().decorate(finalI, j,
-                                realX, (int) Math.round( + finalI + 1), (int) Math.round(x + finalI - 1),
-                                realZ, (int) Math.round(z + j + 1), (int) Math.round(z + j - 1),
+                                realX, Math.round(+finalI + 1), Math.round(x + finalI - 1),
+                                realZ, Math.round(z + j + 1), Math.round(z + j - 1),
                                 output, biome, getDimension().getFluidHeight(), getEngine().getHeight());
                         getSeaFloorDecorator().decorate(finalI, j,
                                 realX, realZ, output, biome, height + 1,
@@ -106,8 +106,8 @@ public class IrisDecorantActuator extends EngineAssignedActuator<BlockData> {
 
                     if (height == getDimension().getFluidHeight()) {
                         getShoreLineDecorator().decorate(finalI, j,
-                                realX, (int) Math.round(x + finalI + 1), (int) Math.round(x + finalI - 1),
-                                realZ, (int) Math.round(z + j + 1), (int) Math.round(z + j - 1),
+                                realX, Math.round(x + finalI + 1), Math.round(x + finalI - 1),
+                                realZ, Math.round(z + j + 1), Math.round(z + j - 1),
                                 output, biome, height, getEngine().getHeight());
                     }
